rename to compression util.
